Added pattern card widget and watch/unwatch abilities
Pattern card for simplicity of adding new cards.
The watch/unwatch functions take an object, the attribute
to watch, the name of the watcher, and the function to
execute when the attribute changes, which takes the
value of the changed attribute as an argument.

The tests on the admin page are useless really,
except for the one that simulates ajax calls to update
the number of new logins in the past 6 hours.
Shows how we can update a variable and have it's value
bound to an element on the page so we don't have to worry about
changing page elements each time a necessary variable updates.